requirements.txt
Django==3.2.8
autopep8==1.6.0
Pillow==8.4.0
django-filter==21.1